refactor: client 패키지명 변경, PaymentClient 인터페이스 service 패키지로 이동
