+++ src/main/java/roomescape/exception/PaymentConfirmClientException.java
@@ -1,6 +1,6 @@
 package roomescape.exception;
 
-import roomescape.client.PaymentErrorResponse;
+import roomescape.infrastructure.PaymentErrorResponse;
 
 public class PaymentConfirmClientException extends RuntimeException {
