+++ src/test/java/roomescape/service/command/PaymentCommandServiceTest.java
@@ -7,17 +7,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import roomescape.client.PaymentClient;
-import roomescape.client.TossErrorResponse;
-import roomescape.client.dto.PaymentConfirmResultDto;
 import roomescape.domain.payment.Payment;
 import roomescape.domain.reservation.Reservation;
 import roomescape.dto.payment.PaymentResponseDto;
 import roomescape.exception.NotFoundException;
 import roomescape.exception.PaymentConfirmClientException;
 import roomescape.exception.PaymentConfirmServerException;
+import roomescape.infrastructure.TossErrorResponse;
+import roomescape.infrastructure.dto.PaymentConfirmResultDto;
 import roomescape.repository.JpaPaymentRepository;
 import roomescape.repository.JpaReservationRepository;
+import roomescape.service.PaymentClient;
 import roomescape.service.dto.PaymentConfirmDto;
 
 import java.util.Optional;
